docker/step_d_finetune/requirements.txt
torch==2.5.1
transformers==4.46.3
peft==0.13.2
accelerate==1.1.1
datasets==3.1.0
Pillow
pyyaml
